requirements.txt
flask
flask-cors
flask-sqlalchemy
imdbpy
tmdbsimple
gunicorn
psycopg2
